refactor(adapters): dissolve runtime.ts into boundary.ts + index.ts
Atrophy the runtime.ts module to nothing. Host injection (Object.assign on
Runtime, ignore list, and react's isReactComponent brand) moves into index.ts;
the host render classes move into a side-effect-free boundary.ts:

- react/boundary.ts: ErrorBoundary + dedupe (renamed from attempt - it
  dedupes fiber-keyed render attempts).
- preact/boundary.ts: ErrorBoundary + the host-ref declare-module; its dedupe
  is a trivial no-op inlined in index.ts.

Runtime.attempt -> Runtime.dedupe, Runtime.boundary -> Runtime.ErrorBoundary.
The isReactComponent brand is a static value (React only reads its presence).
Error-boundary and discarded-render tests migrate to adjacent boundary.test.tsx
in each package. Preact needs no class-detection brand now that core defines
the default Component.prototype.render. PLAN.md updated to match.
