+++ packages/preact/src/jsx-runtime.ts
@@ -4,12 +4,25 @@ import { Fragment, jsx, jsxs } from 'preact/jsx-runtime';
 import { host } from '@expressive/mvc/jsx-runtime';
 
 import type { Component } from '@expressive/mvc';
-import type { ComponentChildren, JSX as PreactJSX } from 'preact';
+import type { ComponentChildren, JSX as PreactJSX, Ref } from 'preact';
 
 declare module '@expressive/mvc/jsx-runtime' {
   interface Host {
     node: ComponentChildren;
     intrinsics: PreactJSX.IntrinsicElements;
+  }
+}
+
+declare module '@expressive/mvc' {
+  namespace Component {
+    interface BaseProps<T extends Component> {
+      /**
+       * Ref which receives the instance of this component.
+       * (Preact JSX does not add `ref` for non-preact classes, so it is
+       * declared here - React infers it from its own class attributes.)
+       */
+      ref?: Ref<T>;
+    }
   }
 }
