+++ packages/react/src/state.ts
@@ -2,7 +2,6 @@ import { State, Context, observer, watch } from '@expressive/mvc';
 import { Runtime, useFactory, useHook, useReady } from './component';
 
 export { Runtime, useFactory, useHook, useReady };
-export { ignore } from './component';
 export { Consumer, Provider, Context, provide } from './context';
 export { use } from './use';
